+++ mongoengine/document.py
@@ -266,7 +266,6 @@ class Document(BaseDocument):
                                                    upsert=True, **write_concern)
                     created = is_new_object(last_error)
 
-
             if cascade is None:
                 cascade = self._meta.get('cascade', False) or cascade_kwargs is not None
 
@@ -441,8 +440,8 @@ class Document(BaseDocument):
 
         .. versionadded:: 0.5
         """
-        import dereference
-        self._data = dereference.DeReference()(self._data, max_depth)
+        DeReference = _import_class('DeReference')
+        DeReference()([self], max_depth + 1)
         return self
 
     def reload(self, max_depth=1):
@@ -451,20 +450,16 @@ class Document(BaseDocument):
         .. versionadded:: 0.1.2
         .. versionchanged:: 0.6  Now chainable
         """
-        id_field = self._meta['id_field']
         obj = self._qs.read_preference(ReadPreference.PRIMARY).filter(
-                **{id_field: self[id_field]}).limit(1).select_related(max_depth=max_depth)
+                **self._object_key).limit(1).select_related(max_depth=max_depth)
 
         if obj:
             obj = obj[0]
         else:
             msg = "Reloaded document has been deleted"
             raise OperationError(msg)
-        for field in self._fields:
+        for field in self._fields_ordered:
             setattr(self, field, self._reload(field, obj[field]))
-        if self._dynamic:
-            for name in self._dynamic_fields.keys():
-                setattr(self, name, self._reload(name, obj._data[name]))
         self._changed_fields = obj._changed_fields
         self._created = False
         return obj